[refactor] remove warning on build

+++ TestShell_Excutor/ShellEraseRange.cpp
@@ -9,16 +9,16 @@ bool EraseRange::execute(CommandInfo cmdInfo)
 		std::swap(start_lba, end_lba);
 		Logger::getInstance()->print(__FUNCTION__, "lba address swap");
 	}
-	if (start_lba < LBA_MIN) start_lba = LBA_MIN;
-	else if (start_lba > LBA_MAX) start_lba = LBA_MAX;
-	if (end_lba < LBA_MIN)  end_lba = LBA_MIN;
-	else if (end_lba > LBA_MAX)     end_lba = LBA_MAX;
+	if (start_lba < (unsigned int)LBA_MIN) start_lba = (unsigned int)LBA_MIN;
+	else if (start_lba > (unsigned int)LBA_MAX) start_lba = (unsigned int)LBA_MAX;
+	if (end_lba < (unsigned int)LBA_MIN)  end_lba = (unsigned int)LBA_MIN;
+	else if (end_lba > (unsigned int)LBA_MAX)     end_lba = (unsigned int)LBA_MAX;
 	int size = end_lba - start_lba + 1;
 
 	int tempsize;
 	if (size > 0) {
 		tempsize = size - 1;
-		if (start_lba + tempsize > LBA_MAX) {
+		if (start_lba + tempsize > (unsigned int)LBA_MAX) {
 			size = LBA_MAX - start_lba + 1;
 		}
 	}
